Add non-generic LinearSelector / LinearMultiSelector / LinearRange so AllViewsTester can discover them
Agent-Logs-Url: https://github.com/gui-cs/Terminal.Gui/sessions/d920e093-379a-4605-8b6b-07b1534379e7
