+++ Terminal.Gui/Views/LinearRange/LinearSelector.cs
@@ -1,117 +1,24 @@
 namespace Terminal.Gui.Views;
 
 /// <summary>
-///     A linear range view that allows selection of a single option from a typed list of options.
+///     Convenience non-generic <see cref="LinearSelector{T}"/> closed over <see cref="string"/>. Allows
+///     designer scenarios (e.g. <c>AllViewsTester</c>) and reflection-based instantiation to discover
+///     and create the view without supplying a type argument.
 /// </summary>
-/// <typeparam name="T">The data type of the options.</typeparam>
 /// <remarks>
 ///     <para>
-///         Exposes the current selection through <see cref="Value"/> as <typeparamref name="T"/>?,
-///         which is <see langword="null"/> when no option is selected.
-///     </para>
-///     <para>
-///         To switch the current selection programmatically, set <see cref="Value"/>. To observe
-///         selection changes, subscribe to <see cref="LinearRangeViewBase{TOption,TValue}.ValueChanged"/>.
+///         To work with non-string option types, use <see cref="LinearSelector{T}"/> directly.
 ///     </para>
 /// </remarks>
-public class LinearSelector<T> : LinearRangeViewBase<T, T>, IDesignable
+public class LinearSelector : LinearSelector<string>
 {
-    private T? _value;
+    /// <summary>Initializes a new instance of <see cref="LinearSelector"/>.</summary>
+    public LinearSelector () { }
 
-    /// <summary>Initializes a new instance of <see cref="LinearSelector{T}"/>.</summary>
-    public LinearSelector () : base (LinearRangeRenderMode.Single) { }
-
-    /// <summary>Initializes a new instance of <see cref="LinearSelector{T}"/>.</summary>
+    /// <summary>Initializes a new instance of <see cref="LinearSelector"/>.</summary>
     /// <param name="options">Initial options.</param>
     /// <param name="orientation">Initial orientation.</param>
-    public LinearSelector (List<T>? options, Orientation orientation = Orientation.Horizontal)
-        : base (options, orientation, LinearRangeRenderMode.Single) { }
-
-    /// <inheritdoc/>
-    public override T? Value
-    {
-        get => _value;
-        set
-        {
-            T? current = _value;
-
-            if (EqualityComparer<T?>.Default.Equals (current, value))
-            {
-                return;
-            }
-
-            if (RaiseValueChanging (current, value))
-            {
-                return;
-            }
-
-            _value = value;
-
-            // Sync indices to match value.
-            if (value is null)
-            {
-                ApplySelectedIndices ([]);
-            }
-            else
-            {
-                int idx = IndexOfData (value);
-
-                if (idx >= 0)
-                {
-                    ApplySelectedIndices ([idx]);
-                }
-                else
-                {
-                    // Value not present among options: clear selection but keep field.
-                    ApplySelectedIndices ([]);
-                }
-            }
-
-            RaiseValueChanged (current, _value);
-        }
-    }
-
-    /// <inheritdoc/>
-    protected override void OnSelectionChanged ()
-    {
-        T? previous = _value;
-        T? newValue = SelectedIndices.Count > 0 ? Options [SelectedIndices [0]].Data : default;
-
-        if (EqualityComparer<T?>.Default.Equals (previous, newValue))
-        {
-            return;
-        }
-
-        _value = newValue;
-        RaiseValueChanged (previous, newValue);
-    }
-
-    /// <summary>
-    ///     Loads demo data suitable for a designer preview: a single-select <see cref="LinearSelector{T}"/>
-    ///     of T-shirt sizes (XS through XXL) with "M" preselected. Only populated when
-    ///     <typeparamref name="T"/> is <see cref="string"/>; for any other type, the view is left untouched
-    ///     and <see langword="false"/> is returned.
-    /// </summary>
-    /// <returns><see langword="true"/> if demo data was loaded.</returns>
-    public virtual bool EnableForDesign ()
-    {
-        if (typeof (T) != typeof (string))
-        {
-            return false;
-        }
-
-        Title = "T-Shirt Size";
-        AssignHotKeys = true;
-        ShowLegends = true;
-
-        string [] sizes = ["XS", "S", "M", "L", "XL", "XXL"];
-
-        Options = sizes.Select (
-                                s => new LinearRangeOption<T> (s, (Rune)s [0], (T)(object)s))
-                       .ToList ();
-
-        Value = (T)(object)"M";
-
-        return true;
-    }
+    public LinearSelector (List<string>? options, Orientation orientation = Orientation.Horizontal)
+        : base (options, orientation)
+    { }
 }
